Sprint Cycle 1:
Main menu screen has been added with options to play the game, view the instructions page and the credits page as well as quitting the application. So far playing the game only moves on to the select difficulty screen and the game itself has not yet been implemented with the various difficulty settings.